+++ dimcli/core/converters.py
@@ -2,27 +2,9 @@ from .utils import dimensions_url
 
 
 # ===========
-# March 4, 2020 : arrived at 'Funder' in CH sheet
+# TESTING - UNSUPPORTED FEATURE
 # ===========
 #
-
-
-# NOTES FOR CH 
-# * we don't have a 'score' value
-# * need to do the extraction from app-metrics instance to have abstracts
-#
-# TODO funder
-# TODO UIDs of supporting grants
-# TODO Supporting Grants (proj number?)
-# TODO Times cited
-# TODO Altmetric
-# TODO Source Linkout
-
-
-
-
-
-# TODO can have a higher class for inheritance
 
 
 class DfConverter(object):
@@ -59,18 +41,44 @@ class DfConverter(object):
         self.column_transformations = {
             # ('new_col_name', 'fun_name')
             'id' : 
-                [('dimensions_url' , 'convert_id_to_url', )],
+                [('Publication ID' , '', ),  # duplicate == keep col
+                 ('Dimensions URL' , 'convert_id_to_url', )],
+            'abstract' : 
+                [('Abstract' , '', )],
+            'doi' : 
+                [('DOI' , '', )],
+            'pmid' : 
+                [('PMID' , '', )],
+            'pmcid' : 
+                [('PMCID' , '', )],
+            'title' : 
+                [('Title' , '', )],
+            'journal.title' : 
+                [('Source title' , '', )],
+            'journal.id' : 
+                [('Source UID' , '', )],
+            'publisher' : 
+                [('Publisher' , '', )],
+            'mesh_terms' : 
+                [('MeSH terms', 'convert_list')],
+            'date' : 
+                [('Publication Date' , '', )],
+            'year' : 
+                [('PubYear' , '', )],
+            'volume' : 
+                [('Volume' , '', )],
+            'issue' : 
+                [('Issue' , '', )],
+            'pages' : 
+                [('Pagination' , '', )],
+            'open_access_categories' : 
+                [('Open Access', 'convert_dict_name')],
+            'type' : 
+                [('Publication Type' , '', )],
             'authors' : 
                 [ ('Authors', 'convert_authors_to_names' ),
                   ('Corresponding Author', 'convert_authors_corresponding' ),
                   ('Authors Affiliations', 'convert_authors_affiliations' )],
-            'author_affiliations' :  # for deprecated fields TODO check corner cases when we have both
-                [ ('Authors2', 'convert_authors_to_names' ),
-                  ('affiliations', 'convert_authors_affiliations' )],      
-            'open_access_categories' : 
-                [('Open Access', 'convert_dict_name')],
-            'mesh_terms' : 
-                [('MeSH terms', 'convert_list')],
             'research_orgs' : 
                 [('Research Organizations - standardized', 'convert_dict_name'),
                  ('GRID IDs', 'convert_dict_ids'),
@@ -78,6 +86,18 @@ class DfConverter(object):
                 #  ('State of Research organization', 'convert_state_name'),
                  ('Country of Research organization', 'convert_country_name'),
                  ],
+            'funders' : 
+                [('Funder', 'convert_dict_name')],
+            'supporting_grant_ids' : 
+                [('UIDs of supporting grants', 'convert_list')],
+            # TODO Supporting Grants (proj number?)
+            'times_cited' : 
+                [('Times cited' , '', )],
+            'altmetric' : 
+                [('Altmetric' , '', )],
+            'linkout' : 
+                [('Source Linkout' , '', )],
+            # NOT USED
             'category_for' : 
                 [('FOR (ANZSRC) Categories', 'convert_dict_name')],
             'category_rcdc' : 
@@ -97,22 +117,28 @@ class DfConverter(object):
         for x in self.columns_original:
             if x in self.column_transformations:
                 for new_col, action in self.column_transformations[x]:
-                    function = getattr(self, action)
-                    # new_col = action + "_new"
-                    if self.verbose: print("Converting ", x, "to", new_col)
-                    if x == "author_affiliationss":
-                        df[new_col] = df[x].fillna("").apply(lambda x: function(x[0]))
+                    if action:
+                        function = getattr(self, action)
+                        # new_col = action + "_new"
+                        if self.verbose: print("Converting ", x, "to", new_col)
+                        if x == "author_affiliationss":
+                            df[new_col] = df[x].fillna("").apply(lambda x: function(x[0]))
+                        else:
+                            df[new_col] = df[x].fillna("").apply(lambda x: function(x))
                     else:
-                        df[new_col] = df[x].fillna("").apply(lambda x: function(x))
+                        df[new_col] = df[x]
 
         if remove_old_columns:
             # drop transformed cols, except 'id'
             cols_to_drop = [*self.column_transformations.keys()]
-            cols_to_drop.remove('id')
+            # cols_to_drop.remove('id') # DEPRECATED
             cols_to_drop = [x for x in cols_to_drop if x in self.columns_original]
             if self.verbose: print("Dropping columns:", cols_to_drop)
             df.drop(columns=cols_to_drop, inplace=True)
         return df
+
+
+    # CONVERSION METHODS
 
 
     def convert_id_to_url(self, idd, ttype=None):