+++ dimcli/tests/quicktest.py
@@ -33,7 +33,7 @@ def main(test_number=1):
         q = f"""search publications 
                 in full_data for "{dsl_escape(covid_q)}" 
                 where year=2020 
-            return publications[id+doi+pmid+pmcid+title+journal+publisher+mesh_terms+date+year+volume+issue+pages+open_access_categories+type+authors+category_for+research_orgs] limit 1000"""
+            return publications[id+doi+pmid+pmcid+title+journal+publisher+mesh_terms+date+year+volume+issue+pages+open_access_categories+type+authors+research_orgs+funders+supporting_grant_ids+times_cited+altmetric+linkout] limit 1000"""
 
         q1 = """search publications for "malaria" return publications[id+authors]"""
 
@@ -43,6 +43,40 @@ def main(test_number=1):
 
         r = DfConverter(df, "publications")
         res = r.simplify_nested_objects()
+
+        order = [   'Publication ID', 
+                    'DOI',
+                    'PMID',
+                    'PMCID',
+                    'Title',
+                    'Abstract',
+                    'Source title',
+                    'Source UID',
+                    'Publisher',
+                    'MeSH terms',
+                    'Publication Date',
+                    'PubYear',
+                    'Volume',
+                    'Issue',
+                    'Pagination',
+                    'Open Access',
+                    'Publication Type',
+                    'Authors',
+                    'Corresponding Author',
+                    'Authors Affiliations',
+                    'Research Organizations - standardized',
+                    'GRID IDs',
+                    'City of Research organization',
+                    'Country of Research organization',
+                    'Funder',
+                    'UIDs of supporting grants',
+                    'Times cited',
+                    'Altmetric',
+                    'Source Linkout',
+                    'Dimensions URL']
+
+        res = res[order]
+
         res.to_csv("test.csv", index=False)